Adding X to the ParseContext is complete useless. Ends with a recursive type.
Adding X to the DynContext ends up with complicated code in the usercode, with multiple typedefs for each context that don't interact too well.

Not adding X to the DynContext but still having it not hidden by DynContext f's up with Context and type ascription (?).

Using multiple lifetimes in a trait definition is plain hell. Turn the heat up if you use them in an associated type.

Not having a definite order for your type variables, or having the same one in different places for different purposes leads to madness.