add combinator track()
breaking?: change combinator transform() to work with ParserError instead of fully generic E. helps type inference.
Also the first E was useless if we assume a parser which returns a ParserError. And hence no need for a WithCode in this place.

add combinator transform_p() where the transformation returns a ParserError instead of a generic E. Don't need an extra error code in this case.

breaking?: Completely different formulation for Context.
FindTracker can be formulated for the input type I, so a single impl of context is possible. Implemented FindTracker for the same input types as before.

This is huge. Now the track() combinator can be expressed. And TrackError doesn't have to bend, it has a very straight forward impl now. And it can be formulated for any I.

Better output formatting.

Test: A few bugs when working with cfg(debug_assertions). Trace and CheckTrace now work too with cfg(debug_assertions).